+++ src/components/NFTcard/nftcard.styles.ts
@@ -1,15 +1,33 @@
 import { makeStyles } from '@material-ui/core/styles'
 
 export const useStyles = makeStyles({
-  root: {
-    maxWidth: 345
+  // root: {
+  //   maxWidth: 810
+  // },
+  card:{
+    height:378,
+    width:310,
+
   },
   media: {
-    height: 175
+    height: 176,
+    marginTop:30,
+    marginLeft:27,
+    marginRight:27,
+   '& img':{
+     width:'100%',
+     height:'100%'
+   }
   },
   price: {
     display: 'flex',
-    justifyContent: 'flex-end'
+    justifyContent: 'flex-start',
+    '& p':{
+    
+
+
+
+    }
   },
   action: {
     display: 'flex',
